+++ packages/react/src/components/List/HierarchyList/HierarchyList.story.jsx
@@ -460,7 +460,7 @@ export const WithLargeNumberOfItems = () => (
   <div style={{ width: 400, height: 400 }}>
     <HierarchyList
       title={text('Title', 'Big List')}
-      isFullHeight={boolean('isFullHeight', true)}
+      isFullHeight={boolean('isFullHeight', false)}
       items={[...Array(1000)].map((_, i) => ({
         id: `item-${i}`,
         content: {
@@ -473,6 +473,7 @@ export const WithLargeNumberOfItems = () => (
       isLoading={boolean('isLoading', false)}
       isLargeRow={boolean('isLargeRow', false)}
       onSelect={action('onSelect')}
+      onListUpdated={action('onListUpdated')}
       hasDeselection={boolean('hasDeselection', true)}
       i18n={object('i18n', {
         searchPlaceHolderText: 'Search',
@@ -483,3 +484,27 @@ export const WithLargeNumberOfItems = () => (
 );
 
 WithLargeNumberOfItems.storyName = 'with large number of items';
+
+export const WithEmptyState = () => (
+  <div style={{ width: 400, height: 400 }}>
+    <HierarchyList
+      title={text('Title', 'Big List')}
+      isFullHeight={boolean('isFullHeight', true)}
+      items={[]}
+      editingStyle={EditingStyle.Single}
+      hasSearch={boolean('hasSearch', true)}
+      pageSize={select('Page Size', ['sm', 'lg', 'xl', undefined], undefined)}
+      isLoading={boolean('isLoading', false)}
+      isLargeRow={boolean('isLargeRow', false)}
+      onSelect={action('onSelect')}
+      hasDeselection={boolean('hasDeselection', true)}
+      i18n={object('i18n', {
+        searchPlaceHolderText: 'Search',
+      })}
+      hasMultiSelect={boolean('hasMultiSelect', false)}
+      emptyState={text('emptyState', '__a custom empty state__')}
+    />
+  </div>
+);
+
+WithEmptyState.storyName = 'with empty state';